+++ packages/app-check-compat/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @firebase/app-check-compat
 
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [[`f36d627af`](https://github.com/firebase/firebase-js-sdk/commit/f36d627af6e1f5ed98e21f9be29f59d2c8c503cb)]:
+  - @firebase/app-check@0.5.12
+
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [[`b12af44a5`](https://github.com/firebase/firebase-js-sdk/commit/b12af44a5c7500e1192d6cc1a4afc4d77efadbaf)]:
+  - @firebase/util@1.6.3
+  - @firebase/app-check@0.5.11
+  - @firebase/component@0.5.17
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [[`efe2000fc`](https://github.com/firebase/firebase-js-sdk/commit/efe2000fc499e2c85c4e5e0fef6741ff3bad2eb0)]:
+  - @firebase/app-check@0.5.10
+  - @firebase/util@1.6.2
+  - @firebase/component@0.5.16
+
 ## 0.2.9
 
 ### Patch Changes