+++ packages/functions-compat/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @firebase/functions-compat
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [[`b12af44a5`](https://github.com/firebase/firebase-js-sdk/commit/b12af44a5c7500e1192d6cc1a4afc4d77efadbaf), [`e673dc808`](https://github.com/firebase/firebase-js-sdk/commit/e673dc808adc14baa499c4ecc31fdb82b1ff0757)]:
+  - @firebase/util@1.6.3
+  - @firebase/functions@0.8.4
+  - @firebase/component@0.5.17
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [[`efe2000fc`](https://github.com/firebase/firebase-js-sdk/commit/efe2000fc499e2c85c4e5e0fef6741ff3bad2eb0)]:
+  - @firebase/util@1.6.2
+  - @firebase/component@0.5.16
+  - @firebase/functions@0.8.3
+
 ## 0.2.2
 
 ### Patch Changes